requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.1
gnureadline==6.3.3
ipython==3.0.0
itsdangerous==0.24
python-consul==0.3.10
requests==2.5.3
six==1.9.0
wsgiref==0.1.2
